pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "zen_europe"
version = "0.0.1"
description="ZEN-europe hosts the european, sector-coupled dataset for ZEN-garden."
readme = "README.md"
requires-python = ">=3.11"
authors = [
  { name="Christoph Funke", email="zen-garden@ethz.ch" },
  { name="Jacob Mannhardt"},
]

dependencies = [
    "zen-creator"
]

[project.optional-dependencies]

docs = [
    "sphinx==7.3.7",
    "myst-parser==3.0.1",
    "sphinx-book-theme==1.1.3",
    "pydata-sphinx-theme==0.15.4",
    "sphinx-reredirects==0.1.4",
    "nbsphinx==0.9.4",
    "nbsphinx-link==1.3.0",
    "sphinxcontrib-mermaid",
    "furo"
]

dev = [
    "pytest",
    "coverage",
    "pytest-xdist",
    "pytest-cov",
    "black==25.12.0",
    "ruff==0.14.13",
    "mypy==1.19.1",
    "types-PyYAML",
    "pandas-stubs",
    "types-geopandas",
    "types-docutils",
    "scipy-stubs"
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra -q"

[tool.black]
line-length = 88
target-version = ["py311", "py312", "py313"]

[tool.ruff]
line-length = 88
fix = false

[tool.ruff.lint]
select = ["E", "F", "I", "B", "D400", "D415"] # add D in future
ignore = []

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.isort]
known-first-party = ["your_package_name"]

[tool.mypy]
#warn_unused_configs = true
#ignore_missing_imports = true
#check_untyped_defs = false
#disallow_untyped_defs = false
exclude = [

]